+++ vue/src/components/Modules/Integrations/PodioIntegrations/List/PodioIntegrationList.vue
@@ -93,7 +93,7 @@ function onEditItemBtnClicked(item: PodioIntegration) {
 
 function deleteItem(item: PodioIntegration) {
     bus.emit('confirm', {
-        body: `Do you want to delete <strong>${item.name}</strong>?`,
+        body: `Do you want to delete "${item.name}"?`,
         confirmIcon: 'fa fa-trash',
         confirmColor: 'red',