+++ vue/src/components/Modules/Setup/ContainerImages/List/ContainerImageList.vue
@@ -118,7 +118,7 @@ function onDuplicateItemBtnClicked(item: ContainerImage) {
 
 function deleteItem(item: ContainerImage) {
     bus.emit('confirm', {
-        body: `Do you want to delete <strong>${item.name}</strong>?`,
+        body: `Do you want to delete "${item.name}"?`,
         confirmIcon: 'fa fa-trash',
         confirmColor: 'red',