+++ vue/src/components/Modules/Setup/DeploymentPackages/List/DeploymentPackageList.vue
@@ -91,7 +91,7 @@ function createItem() {
 
 function onDeleteItemBtnClicked(item: DeploymentPackage) {
     bus.emit('confirm', {
-        body: `Do you want to delete <strong>${item.name}</strong>?`,
+        body: `Do you want to delete "${item.name}"?`,
         confirmIcon: 'fa fa-trash',
         confirmColor: 'red',