+++ vue/src/components/Modules/Setup/DeploymentSpecifications/List/DeploymentSpecificationList.vue
@@ -112,7 +112,7 @@ function createItem(type: string) {
 
 function onDeleteItemBtnClicked(item: DeploymentSpecification) {
     bus.emit('confirm', {
-        body: `Do you want to delete <strong>${item.name}</strong>?`,
+        body: `Do you want to delete "${item.name}"?`,
         confirmIcon: 'fa fa-trash',
         confirmColor: 'red',