+++ vue/src/components/Modules/Setup/Deployments/List/DeploymentList.vue
@@ -156,7 +156,7 @@ function onCreateItemBtnClicked(deploymentSpec: DeploymentSpecification) {
 
 function onDeleteItemBtnClicked(item: Deployment) {
     bus.emit('confirm', {
-        body: `Do you want to delete <strong>${item.name}</strong>?`,
+        body: `Do you want to delete "${item.name}"?`,
         confirmIcon: 'fa fa-trash',
         confirmColor: 'red',