+++ vue/src/components/Modules/Setup/EmailServices/List/EmailServiceList.vue
@@ -106,7 +106,7 @@ function onDuplicateItemBtnClicked(item: EmailService) {
 
 function onDeleteItemBtnClicked(item: EmailService) {
     bus.emit('confirm', {
-        body: `Do you want to delete <strong>${item.name}</strong>?`,
+        body: `Do you want to delete "${item.name}"?`,
         confirmIcon: 'fa fa-trash',
         confirmColor: 'red',