+++ vue/src/components/Modules/Users/List/UserList.vue
@@ -96,7 +96,7 @@ function editItem(item: User) {
 
 function deleteItem(item: User) {
     bus.emit('confirm', {
-        body: `Do you want to delete <strong>${item.name}</strong>?`,
+        body: `Do you want to delete "${item.name}"?`,
         confirmIcon: 'fa fa-trash',
         confirmColor: 'red',